app/flask_app/requirements.txt
flask
python-dotenv
environs
psycopg2-binary
Flask-SQLAlchemy
